requirements.txt
torch>=1.7.0
torchvision>=0.8.0
transformers>=4.5.0
pillow>=8.0.0
matplotlib>=3.3.0
numpy>=1.19.0
gradio>=3.0.0
requests>=2.25.0